fix: move shadcn components from literal @/ to client/components/ui
shadcn cli wrote to literal @/ directory instead of resolving alias.
move files to correct location and delete @/ directory.
